ambient-assistant-mac/requirements.txt
# Core dependencies
PyQt6>=6.6.0
PySide6>=6.6.0
pyobjc-core>=10.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
pyobjc-framework-Vision>=10.0
python-dotenv>=1.0.0
pillow>=10.0.0
pytesseract>=0.3.10
numpy>=1.24.0
opencv-python>=4.8.0
pynput>=1.7.6
requests>=2.31.0
markdown>=3.5.0

# Django for database and API
Django>=4.2.0
djangorestframework>=3.14.0

# LLM dependencies
openai>=1.3.0
anthropic>=0.8.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-anthropic>=0.1.1
langchain-community>=0.0.10
langchain-core>=0.1.10
langgraph>=0.0.20
crewai>=0.16.0
sentence-transformers>=2.2.2
tiktoken>=0.5.0

# Utilities
pyperclip>=1.8.2
psutil>=5.9.0
appdirs>=1.4.4
watchdog>=3.0.0
rich>=13.5.0
pyinstaller>=6.0.0  # For building distributable
pytest>=7.4.0  # For testing

# Optional: Audio processing dependencies
pyaudio>=0.2.13
sounddevice>=0.4.6
librosa>=0.10.1

# Optional: Document processing
# For handling different document formats if needed
pypdf>=3.17.0
docx2txt>=0.8
beautifulsoup4>=4.12.0
